+++ src/graphql/index.ts
@@ -9,14 +9,14 @@ import {
   resolvers as scalarResolvers,
 } from 'graphql-scalars';
 import { portfolioResolvers, portfolioTypeDefs } from './portfolio';
-import { holdingsTypeDefs } from './holdings/typeDefs';
-import { holdingsResolvers } from './holdings/resolvers';
+// import { holdingsTypeDefs } from './holdings/typeDefs';
+// import { holdingsResolvers } from './holdings/resolvers';
 import { transactionsResolvers, transactionsTypeDefs } from './transactions';
 import { fundsResolvers, fundsTypeDefs } from './funds';
 // import { documentsResolvers, documentsTypedefs } from './documents';
 import { marketNewsResolvers, marketNewsTypeDefs } from './news';
 import { calculatorsResolvers, calculatorsTypeDefs } from './calculators';
-import { goalsResolvers, goalsTypeDefs } from './goals';
+// import { goalsResolvers, goalsTypeDefs } from './goals';
 import { chatBotResolvers, chatBotTypeDefs } from './chatBot';
 import { accountResolvers, accountTypeDefs } from './accounts';
 import { lookupsResolvers, lookupsTypeDefs } from './lookups';
@@ -27,13 +27,13 @@ const typeDefs = [
   userTypeDefs,
   scalarTypeDefs,
   portfolioTypeDefs,
-  holdingsTypeDefs,
+  // holdingsTypeDefs,
   transactionsTypeDefs,
   fundsTypeDefs,
   // documentsTypedefs,
   marketNewsTypeDefs,
   calculatorsTypeDefs,
-  goalsTypeDefs,
+  // goalsTypeDefs,
   chatBotTypeDefs,
   accountTypeDefs,
   lookupsTypeDefs,
@@ -45,13 +45,13 @@ const resolvers = [
   UserResolvers,
   scalarResolvers,
   portfolioResolvers,
-  holdingsResolvers,
+  // holdingsResolvers,
   transactionsResolvers,
   fundsResolvers,
   // documentsResolvers,
   marketNewsResolvers,
   calculatorsResolvers,
-  goalsResolvers,
+  // goalsResolvers,
   chatBotResolvers,
   accountResolvers,
   lookupsResolvers,
