+++ src/graphql/shield/index.ts
@@ -4,9 +4,9 @@ import { userShield } from './user';
 import { portfolioShield } from './portfolio';
 import { performanceShield } from './performance';
 import { fundsShield } from './funds';
-import { holdingsShield } from './holdings';
+// import { holdingsShield } from './holdings';
 import { transactionsShield } from './transactions';
-import { goalsShield } from './goals';
+// import { goalsShield } from './goals';
 import { accountShield } from './account';
 
 export const permissions = shield(
@@ -16,9 +16,7 @@ export const permissions = shield(
       ...portfolioShield.Query,
       ...performanceShield.Query,
       ...fundsShield.Query,
-      ...holdingsShield.Query,
       ...transactionsShield.Query,
-      ...goalsShield.Query,
       ...accountShield.Query,
     },
     Mutation: {
@@ -27,7 +25,6 @@ export const permissions = shield(
       ...performanceShield.Mutation,
       ...fundsShield.Mutation,
       ...transactionsShield.Mutation,
-      ...goalsShield.Mutation,
       ...accountShield.Mutation,
     },
   },
