+++ src/services/users/index.ts
@@ -1,3 +1,2 @@
 export * from './user';
-export * from './kyc';
 export * from './kyc.service';
